+++ eslint.config.js
@@ -1,4 +1,5 @@
 import svelte_config from '@sveltejs/eslint-config';
+import noRuntimeToExportsImports from './.eslint/no-runtime-to-exports-imports.js';
 
 /** @type {import('eslint').Linter.Config[]} */
 export default [
@@ -9,10 +10,25 @@ export default [
 		}
 	},
 	{
+		files: ['packages/kit/src/runtime/**/*.js'],
+		plugins: {
+			'kit-custom': {
+				rules: {
+					'no-runtime-to-exports-imports': noRuntimeToExportsImports
+				}
+			}
+		},
+		rules: {
+			'kit-custom/no-runtime-to-exports-imports': 'error'
+		}
+	},
+	{
 		ignores: [
 			'**/.svelte-kit',
+			'**/.wrangler',
 			'**/test-results',
 			'**/build',
+			'**/dist',
 			'**/.custom-out-dir',
 			'packages/adapter-*/files',
 			'packages/kit/src/core/config/fixtures/multiple', // dir contains svelte config with multiple extensions tripping eslint