+++ packages/kit/test/apps/basics/src/routes/remote/form/submitter/form.remote.ts
@@ -3,11 +3,9 @@ import * as v from 'valibot';
 
 export const my_form = form(
 	v.object({
-		foo: v.picklist(['a', 'b', 'c']),
-		bar: v.picklist(['d', 'e', 'f']),
-		button: v.literal('submitter')
+		submitter: v.string()
 	}),
 	async (data) => {
-		console.log(data);
+		return data.submitter;
 	}
 );