+++ packages/package/test/fixtures/resolve-alias/expected/baz.js
@@ -1,2 +1,0 @@
-export const baz = { baz: 'baz' };
-export default baz;